[Security AI] Move `security-ai-prompt` to package (#208787)
(cherry picked from commit b0a72e2b9086728d98f4176d1692741f4df62333)

# Conflicts:
#	.github/CODEOWNERS
