+++ tsconfig.base.json
@@ -1624,6 +1624,8 @@
       "@kbn/search-types/*": ["src/platform/packages/shared/kbn-search-types/*"],
       "@kbn/searchprofiler-plugin": ["x-pack/platform/plugins/shared/searchprofiler"],
       "@kbn/searchprofiler-plugin/*": ["x-pack/platform/plugins/shared/searchprofiler/*"],
+      "@kbn/security-ai-prompts": ["x-pack/solutions/security/packages/security-ai-prompts"],
+      "@kbn/security-ai-prompts/*": ["x-pack/solutions/security/packages/security-ai-prompts/*"],
       "@kbn/security-api-integration-helpers": ["x-pack/test/security_api_integration/packages/helpers"],
       "@kbn/security-api-integration-helpers/*": ["x-pack/test/security_api_integration/packages/helpers/*"],
       "@kbn/security-api-key-management": ["x-pack/platform/packages/shared/security/api_key_management"],
