+++ x-pack/solutions/security/packages/security-ai-prompts/README.md
@@ -1,0 +1,3 @@
+# @kbn/security-ai-prompts
+
+Utility library for Security AI Prompt management.
