+++ x-pack/solutions/security/packages/security-ai-prompts/src/utils.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const getProviderFromActionTypeId = (actionTypeId: string): string | undefined => {
+  const providerDictionary: Record<string, string> = {
+    [`.gen-ai`]: `openai`,
+    [`.bedrock`]: `bedrock`,
+    [`.gemini`]: `gemini`,
+    [`.inference`]: `inference`,
+  };
+  return providerDictionary[actionTypeId];
+};
