+++ x-pack/solutions/security/plugins/elastic_assistant/common/constants.ts
@@ -31,6 +31,3 @@ export const CAPABILITIES = `${BASE_PATH}/capabilities`;
  Licensing requirements
  */
 export const MINIMUM_AI_ASSISTANT_LICENSE = 'enterprise' as const;
-
-// Saved Objects
-export const promptSavedObjectType = 'security-ai-prompt';
