+++ x-pack/solutions/security/plugins/elastic_assistant/tsconfig.json
@@ -54,7 +54,7 @@
     "@kbn/llm-tasks-plugin",
     "@kbn/product-doc-base-plugin",
     "@kbn/core-saved-objects-api-server-mocks",
-    "@kbn/inference-common"
+    "@kbn/security-ai-prompts"
   ],
   "exclude": [
     "target/**/*",
