+++ x-pack/solutions/security/plugins/security_solution/server/saved_objects.ts
@@ -7,6 +7,7 @@
 
 import type { CoreSetup } from '@kbn/core/server';
 
+import { promptType } from '@kbn/security-ai-prompts';
 import { protectionUpdatesNoteType } from './endpoint/lib/protection_updates_note/saved_object_mappings';
 import { noteType, pinnedEventType, timelineType } from './lib/timeline/saved_object_mappings';
 // eslint-disable-next-line no-restricted-imports
@@ -29,6 +30,7 @@ const types = [
   riskEngineConfigurationType,
   entityEngineDescriptorType,
   protectionUpdatesNoteType,
+  promptType,
 ];
 
 export const savedObjectTypes = types.map((type) => type.name);
