+++ x-pack/solutions/security/plugins/security_solution/tsconfig.json
@@ -236,6 +236,7 @@
     "@kbn/llm-tasks-plugin",
     "@kbn/charts-theme",
     "@kbn/product-doc-base-plugin",
-    "@kbn/shared-ux-error-boundary"
+    "@kbn/shared-ux-error-boundary",
+    "@kbn/security-ai-prompts",
   ]
 }
