+++ x-pack/test_serverless/api_integration/test_suites/security/platform_security/authorization.ts
@@ -377,6 +377,18 @@ export default function ({ getService }: FtrProviderContext) {
                 "saved_object:policy-settings-protection-updates-note/delete",
                 "saved_object:policy-settings-protection-updates-note/bulk_delete",
                 "saved_object:policy-settings-protection-updates-note/share_to_space",
+                "saved_object:security-ai-prompt/bulk_get",
+                "saved_object:security-ai-prompt/get",
+                "saved_object:security-ai-prompt/find",
+                "saved_object:security-ai-prompt/open_point_in_time",
+                "saved_object:security-ai-prompt/close_point_in_time",
+                "saved_object:security-ai-prompt/create",
+                "saved_object:security-ai-prompt/bulk_create",
+                "saved_object:security-ai-prompt/update",
+                "saved_object:security-ai-prompt/bulk_update",
+                "saved_object:security-ai-prompt/delete",
+                "saved_object:security-ai-prompt/bulk_delete",
+                "saved_object:security-ai-prompt/share_to_space",
                 "saved_object:csp_rule/bulk_get",
                 "saved_object:csp_rule/get",
                 "saved_object:csp_rule/find",
@@ -1223,6 +1235,18 @@ export default function ({ getService }: FtrProviderContext) {
                 "saved_object:policy-settings-protection-updates-note/delete",
                 "saved_object:policy-settings-protection-updates-note/bulk_delete",
                 "saved_object:policy-settings-protection-updates-note/share_to_space",
+                "saved_object:security-ai-prompt/bulk_get",
+                "saved_object:security-ai-prompt/get",
+                "saved_object:security-ai-prompt/find",
+                "saved_object:security-ai-prompt/open_point_in_time",
+                "saved_object:security-ai-prompt/close_point_in_time",
+                "saved_object:security-ai-prompt/create",
+                "saved_object:security-ai-prompt/bulk_create",
+                "saved_object:security-ai-prompt/update",
+                "saved_object:security-ai-prompt/bulk_update",
+                "saved_object:security-ai-prompt/delete",
+                "saved_object:security-ai-prompt/bulk_delete",
+                "saved_object:security-ai-prompt/share_to_space",
                 "saved_object:csp_rule/bulk_get",
                 "saved_object:csp_rule/get",
                 "saved_object:csp_rule/find",
@@ -1840,6 +1864,11 @@ export default function ({ getService }: FtrProviderContext) {
                 "saved_object:policy-settings-protection-updates-note/find",
                 "saved_object:policy-settings-protection-updates-note/open_point_in_time",
                 "saved_object:policy-settings-protection-updates-note/close_point_in_time",
+                "saved_object:security-ai-prompt/bulk_get",
+                "saved_object:security-ai-prompt/get",
+                "saved_object:security-ai-prompt/find",
+                "saved_object:security-ai-prompt/open_point_in_time",
+                "saved_object:security-ai-prompt/close_point_in_time",
                 "saved_object:csp_rule/bulk_get",
                 "saved_object:csp_rule/get",
                 "saved_object:csp_rule/find",
@@ -2206,6 +2235,11 @@ export default function ({ getService }: FtrProviderContext) {
                 "saved_object:policy-settings-protection-updates-note/find",
                 "saved_object:policy-settings-protection-updates-note/open_point_in_time",
                 "saved_object:policy-settings-protection-updates-note/close_point_in_time",
+                "saved_object:security-ai-prompt/bulk_get",
+                "saved_object:security-ai-prompt/get",
+                "saved_object:security-ai-prompt/find",
+                "saved_object:security-ai-prompt/open_point_in_time",
+                "saved_object:security-ai-prompt/close_point_in_time",
                 "saved_object:csp_rule/bulk_get",
                 "saved_object:csp_rule/get",
                 "saved_object:csp_rule/find",
